+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Spell check scopes are configurable: you can specify whether to check or ignore spelling inside comments, math, latex-image, code, etc.
 - Support for CLI 2.5+ schema locations.
 
+### Changed
+
+- Improved feedback messages when running PreTeXt commands.
+
 ## [0.19.0] - 2024-07-06
 
 ### Added